Fix the bug on the Infinite Roulette mode
When the user plays the Infinite Roulette, the program abruptly end the
program early. It's because of how the number of mines is
genereated in this mode, the rows and columns are generated in random
having a limit of 9 for rows and 20 for columns but since rand() returns
number starting from zero, if in any case atleast one (either rows or
columns) have a value of zero, then the in generating the number of
mines which depends on the board size (rows * columns) and also I want
the mines to be at most half of the board size, which will give rand() a
divisor of 0 which results in undefined behaviour of the program.
So to fix it what i made is that i added 3 after the number of rows and
columns are generated, and also made sure that the loop ends if the rows
and cols are higher then the minumum size (3) and lower than the maximum
size which is (rows -> 9, columns -> 20), and also the way the number of
mines is generated is still the same it will generate atleast 3 bombs
per board generated.