Fix the leaderboard print format and how the program read data in the txt database.
I fix the leaderboard format when the number of players in the ranking
goes over 1 digit since the table adjust the spaces causing the row of
the table to have more lenght, I fix the spacing so that even if the
leaderboard have more players in the leaderboard the spacing would still
be the same.

Instead of creating one Leaderboard entry and every time the program
read the data from the file it would just change the value of the same
LeaderboardEntry and passed it in the vector, I make it so that it will
create new Leaderboard object everytime it get data so that what i
passed in the vector is different objects of the Leaderboard structure.