Fix the bug in the leaderboad where it prints players with emtpy names
I found out that this is because i use the .eof() method of fstream, it
checks until the blankline of the file which made it also read the
blankline resulting on adding the empty player names when reading the
file into the leaderboard vector. Fix it by checking if the player name
it returns is empty, if it is then we just end the file reading.